aec-data-model-assistant/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozenlist==1.5.0
gql==3.5.0
graphql-core==3.2.6
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
langgraph==0.2.73
langgraph-checkpoint==2.0.15
langgraph-sdk==0.1.51
langsmith==0.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.63.2
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0
yarl==1.18.3
zstandard==0.23.0
